requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
python-dotenv
openai
tqdm
pydantic